pyproject.toml
[project]
name = "cinemate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dateparser>=1.2.2",
    "fastmcp>=2.13.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "httpx>=0.28.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pytz",
    "tzlocal",
]
